requirements.txt
python-swiftclient==3.2.0
python-keystoneclient==3.9.0
Flask==0.12
cf-deployment-tracker==1.0.3
pyDes==2.0.1